Remove RuntimeEvaluatorExtractors
